Sanitize new worktrees before use.
When using `workman new` with `git worktree`, creating a new working
directory will create a new worktree, which by default is on the same
branch and has the same untracked changes as the root worktree.

Before proceeding to fetch upstream's default branch, we need to clean
and switch to the current default branch.

Signed-off-by: David Wood <david@davidtw.co>